[chained|proxyimpl] Move protocol implementations to proxyimpl
