+++ common/nopcloser.go
@@ -1,0 +1,5 @@
+package common
+
+type NopCloser struct{}
+
+func (NopCloser) Close() {}
