+++ proxyimpl/multipath.go
@@ -1,4 +1,4 @@
-package chained
+package proxyimpl
 
 import (
 	"context"
@@ -7,14 +7,13 @@ import (
 
 	"github.com/getlantern/common/config"
 	"github.com/getlantern/errors"
-	"github.com/getlantern/flashlight/balancer"
 	"github.com/getlantern/flashlight/common"
 	"github.com/getlantern/flashlight/ops"
 	"github.com/getlantern/multipath"
 )
 
 type mpDialerAdapter struct {
-	impl  proxyImpl
+	impl  ProxyImpl
 	label string
 }
 
@@ -29,46 +28,43 @@ func (d *mpDialerAdapter) DialContext(ctx context.Context) (net.Conn, error) {
 		op = ops.Begin("dial_subflow")
 	}
 	defer op.End()
-	return d.impl.dialServer(op, ctx)
+	return d.impl.DialServer(op, ctx)
 }
 
 func (d *mpDialerAdapter) Label() string {
 	return d.label
 }
 
-type multipathImpl struct {
-	nopCloser
+type MultipathImpl struct {
+	common.NopCloser
 	dialer multipath.Dialer
 }
 
-func (impl *multipathImpl) dialServer(op *ops.Op, ctx context.Context) (net.Conn, error) {
+func (impl *MultipathImpl) DialServer(op *ops.Op, ctx context.Context) (net.Conn, error) {
 	return impl.dialer.DialContext(context.WithValue(ctx, "op", op))
 }
 
-func (impl *multipathImpl) FormatStats() []string {
+func (impl *MultipathImpl) FormatStats() []string {
 	return impl.dialer.(multipath.Stats).FormatStats()
 }
 
-func CreateMPDialer(configDir, endpoint string, ss map[string]*config.ProxyConfig, uc common.UserConfig) (balancer.Dialer, error) {
+func NewMultipathProxyImpl(
+	configDir, endpoint string,
+	ss map[string]*config.ProxyConfig,
+	uc common.UserConfig,
+	extractProxyConfigParamsFunc func(*config.ProxyConfig) (addr string, transport string, network string, err error),
+	defaultReportDialCore ReportDialCoreFn,
+) (ProxyImpl, error) {
 	if len(ss) < 1 {
 		return nil, errors.New("no dialers")
 	}
-	var p *proxy
-	var err error
 	var dialers []multipath.Dialer
 	for name, s := range ss {
-		if p == nil {
-			// Note: we pass the first server info to newProxy for the attributes shared by all paths
-			p, err = newProxy(endpoint, endpoint+":0", "multipath", "multipath", s, uc)
-			if err != nil {
-				return nil, err
-			}
-		}
-		addr, transport, _, err := extractParams(s)
+		addr, transport, _, err := extractProxyConfigParamsFunc(s)
 		if err != nil {
 			return nil, err
 		}
-		impl, err := createImpl(configDir, name, addr, transport, s, uc, p.reportDialCore)
+		impl, err := New(configDir, name, addr, transport, s, uc, defaultReportDialCore)
 		if err != nil {
 			log.Errorf("failed to add %v to %v, continuing: %v", s.Addr, name, err)
 			continue
@@ -79,11 +75,10 @@ func CreateMPDialer(configDir, endpoint string, ss map[string]*config.ProxyConfi
 	if len(dialers) == 0 {
 		return nil, errors.New("no subflow dialer")
 	}
-	p.impl = &multipathImpl{dialer: multipath.NewDialer(endpoint, dialers)}
-	return p, nil
+	return &MultipathImpl{dialer: multipath.NewDialer(endpoint, dialers)}, nil
 }
 
-func groupByMultipathEndpoint(proxies map[string]*config.ProxyConfig) map[string]map[string]*config.ProxyConfig {
+func GroupByMultipathEndpoint(proxies map[string]*config.ProxyConfig) map[string]map[string]*config.ProxyConfig {
 	groups := make(map[string]map[string]*config.ProxyConfig)
 	for name, s := range proxies {
 		group, exists := groups[s.MultipathEndpoint]
