+++ proxyimpl/obfs4.go
@@ -1,4 +1,4 @@
-package chained
+package proxyimpl
 
 import (
 	"context"
@@ -7,20 +7,21 @@ import (
 
 	pt "git.torproject.org/pluggable-transports/goptlib.git"
 	"github.com/getlantern/common/config"
+	"github.com/getlantern/flashlight/common"
 	"github.com/getlantern/flashlight/ops"
 	"gitlab.com/yawning/obfs4.git/transports/base"
 	"gitlab.com/yawning/obfs4.git/transports/obfs4"
 )
 
 type obfs4Impl struct {
-	nopCloser
+	common.NopCloser
 	dialCore coreDialer
 	addr     string
 	cf       base.ClientFactory
 	args     interface{}
 }
 
-func newOBFS4Impl(name, addr string, pc *config.ProxyConfig, dialCore coreDialer) (proxyImpl, error) {
+func newOBFS4Impl(name, addr string, pc *config.ProxyConfig, dialCore coreDialer) (ProxyImpl, error) {
 	if pc.Cert == "" {
 		return nil, fmt.Errorf("No Cert configured for obfs4 server, can't connect")
 	}
@@ -47,7 +48,7 @@ func newOBFS4Impl(name, addr string, pc *config.ProxyConfig, dialCore coreDialer
 	}, nil
 }
 
-func (impl *obfs4Impl) dialServer(op *ops.Op, ctx context.Context) (net.Conn, error) {
+func (impl *obfs4Impl) DialServer(op *ops.Op, ctx context.Context) (net.Conn, error) {
 	dial := func(network, address string) (net.Conn, error) {
 		// We know for sure the network and address are the same as what
 		// the inner DailServer uses.
