+++ proxyimpl/preconnecting_dialer_test.go
@@ -1,4 +1,4 @@
-package chained
+package proxyimpl
 
 import (
 	"context"
@@ -28,11 +28,11 @@ func (conn *testConn) Close() error {
 	return nil
 }
 
-func newEmptyDialer() proxyImpl {
+func newEmptyDialer() ProxyImpl {
 	iod := int64(0)
 	ipc := int64(0)
 
-	return &testImpl{d: func(ctx context.Context) (net.Conn, error) {
+	return &TestImpl{DialServerFunc: func(ctx context.Context) (net.Conn, error) {
 		var id string
 		if ctx == nil {
 			if atomic.LoadInt64(&iod) >= maxSuccessfulDials {
@@ -51,9 +51,9 @@ func newEmptyDialer() proxyImpl {
 
 func TestPreconnecting(t *testing.T) {
 	pd := newPreconnectingDialer("testPreconnecting", 2, 10*time.Second, newEmptyDialer())
-	defer pd.close()
+	defer pd.Close()
 
-	conn, err := pd.dialServer(nil, nil)
+	conn, err := pd.DialServer(nil, nil)
 	if !assert.NoError(t, err) {
 		return
 	}
@@ -62,7 +62,7 @@ func TestPreconnecting(t *testing.T) {
 	for i := 1; i <= maxSuccessfulPreconnects; i++ {
 		time.Sleep(25 * time.Millisecond)
 		assert.Equal(t, 1, pd.numPreconnected())
-		conn, err = pd.dialServer(nil, nil)
+		conn, err = pd.DialServer(nil, nil)
 		if !assert.NoError(t, err) {
 			return
 		}
@@ -72,30 +72,30 @@ func TestPreconnecting(t *testing.T) {
 	for i := 1; i < maxSuccessfulDials; i++ {
 		t.Log(i)
 		time.Sleep(25 * time.Millisecond)
-		conn, err = pd.dialServer(nil, nil)
+		conn, err = pd.DialServer(nil, nil)
 		if !assert.NoError(t, err) {
 			return
 		}
 		assert.Equal(t, fmt.Sprintf("ondemand-%d", i+1), conn.(*testConn).id, "Should have gotten on demand dialed connection after preconnects started failing")
 	}
 
-	conn, err = pd.dialServer(nil, nil)
+	conn, err = pd.DialServer(nil, nil)
 	assert.Error(t, err, "Should have failed to dial after on demand dials started failing")
 }
 
 func TestPreconnectingTimeout(t *testing.T) {
 	expiration := 250 * time.Millisecond
 	pd := newPreconnectingDialer("testPreconnectingTimeout", 2, expiration, newEmptyDialer())
-	defer pd.close()
+	defer pd.Close()
 
 	for i := 1; i <= maxSuccessfulPreconnects; i++ {
 		// dial to fill up pool
-		pd.dialServer(nil, nil)
+		pd.DialServer(nil, nil)
 	}
 
 	// wait for preconnected connections to expire
 	time.Sleep(expiration * 2)
-	conn, err := pd.dialServer(nil, nil)
+	conn, err := pd.DialServer(nil, nil)
 	if assert.NoError(t, err) {
 		assert.NotContains(t, conn.(*testConn).id, "preconnected", "After all preconnected connections expired, we should dial on demand")
 	}
@@ -107,7 +107,7 @@ func TestPreconnectingClose(t *testing.T) {
 
 	for i := 1; i <= maxSuccessfulPreconnects; i++ {
 		// dial to fill up pool
-		pd.dialServer(nil, nil)
+		pd.DialServer(nil, nil)
 	}
 
 	// wait a little for preconnecting to finish
@@ -115,14 +115,14 @@ func TestPreconnectingClose(t *testing.T) {
 	assert.True(t, len(pd.pool) >= 2, "should have at least 2 preconnections")
 	assert.True(t, len(pd.pool) <= 4, "should have no more than 4 preconnections")
 
-	pd.close()
+	pd.Close()
 	// wait a little for closing to happen
 	time.Sleep(25 * time.Millisecond)
 	assert.Empty(t, pd.pool)
 
 	// wait for preconnected connections to expire
 	time.Sleep(expiration * 2)
-	conn, err := pd.dialServer(nil, nil)
+	conn, err := pd.DialServer(nil, nil)
 	if assert.NoError(t, err) {
 		assert.NotContains(t, conn.(*testConn).id, "preconnected", "After all preconnected connections expired, we should dial on demand")
 	}
