+++ proxyimpl/session_caching_test.go
@@ -1,4 +1,4 @@
-package chained
+package proxyimpl
 
 import (
 	"testing"
