+++ proxyimpl/tls_config_ios.go
@@ -1,6 +1,6 @@
 //go:build ios
 
-package chained
+package proxyimpl
 
 import (
 	"context"
